Do not touch entire cache on USB SDMA on RPi 1
Until now before and after slave DMA operations to/from the USB host
controller the entire data cache was cleaned and invalidated on the
Raspberry Pi 1. These cache maintenance operations may be interrupted
and have to be restarted in this case on ARMv6 which was not done.
This led to some strange system behaviour under heavy USB load and
probably to issue #10. Now only the cache range of the SDMA data
buffer will be touched as it was already the case on the Raspberry
Pi 2. This should also increase USB performance a bit on the Pi 1.

Until now the IRQ handler was writing to the IRQ pending register to
reset the IRQ pending bit of the handled IRQ. This does not work and
does not need to be done. The pending bit is reset automatically when
the interrupt condition is cleared on the slave device.