Serial baud rate may be up to 3 Mbps now
init_uart_clock must be set in config.txt to a value above 3000000
(e.g. 16000000) to use this with a serial converter which supports
such speeds.

Kernel listing in kernel.lst improved. C++ names are demangled now
in the listing and data in the text segment is identified and not
translated into instructions of no sense.

Suggested by @jsanchezv