Initial commit of development version "26alpha"
When building with "RASPPI = 2" the resulting kernel image is now
kernel7.img which runs on both RPi 2 and 3. When building with
"RASPPI = 3" the the resulting kernel7.img is optimized for RPi 3.

There is an instance of class CMachineInfo created early in sysinit.cpp
which requests the hardware revision number, decodes it and caches the
results. Further instances of this class only copy these values. This
class has been extended by methods for getting info about clocks and
peripheral devices.

The ARMv7 cache maintenance routines are replaced by optimized versions
which work for both RPi 2 and 3 without configuration.

CGPIOPin::CGPIOPin() can get a "virtual" pin number now, which will be
mapped onto a physical number depending on the RPi model.

Property mailbox calls use the coherent region on all RPi models now.

There is only one Rules.mk file in project root required any more.
The files in app/ and sample/ are provided for compatibility.