Two fixes
CBcmMailBox::Write() did test the wrong register for checking status.
Because of the FIFO and the nature of communication the full bit
was never set, so this has not been a serious problem.

CBcmFrameBuffer::Initialize() did not work on RPi 2 with -O3.
This problem came in with the new cache maintenance routines
in version 26alpha.