Using CUSBKeyboardDevice, CCPUThrottle made easier
Both classes require to repeatedly call methods to update
the LED status resp. the CPU clock from non-IRQ context. For
the LED status it was required to call GetLEDStatus() and SetLEDs()
with some logic and for the CPU clock it had to be ensured that
SetOnTemperature() gets called in an interval of 2 to 5 seconds.
Now one only has to call the following methods as often as wanted
(the previous methods still work):

CUSBKeyboardDevice::UpdateLEDs()
CCPUThrottle::Update()

sample/08-usbkeyboard updated