FIXED: TCP connection may lose data before FIN
If a FIN arrived in a TCP connection together with out-of-sequence data,
because one or more TCP segments got lost, this FIN had been handled and
the connection was closed. This could lead to a lose of data at the end
of a document transferred using TCP. The FIN will be ignored now, if it
is out-of-sequence.