Improved periodic USB delays with USE_USB_SOF_INTR
The USB HCI driver did limit USB interrupt endpoint intervals to 20ms
until now, because of the resolution of the used kernel timers and to
reduce the IRQ rate. With the USE_USB_SOF_INTR option it is possible
to use the USB transaction queue to implement periodic USB delays.
This allows intervals down to 1ms now.