Build option STDLIB_SUPPORT added
If "STDLIB_SUPPORT=1" is defined in Config.mk (or Rules.mk) Circle will
be built with the following peculiarities:

* C++ exceptions can be used.
* C++ Runtime type info is available.
* Applications will be linked using libstdc++.a.
* Applications will be linked using libgcc_eh.a (if available).
* An instance of the class CMemorySystem will be created in sysinit()
  before calling the static constructors. This must not be done
  in CKernel then.
* For some utility functions a specific name prefix is used to prevent
  "multiple definition" errors from the linker.
* Some externals from a stdlib implementation (e.g. newlib) are needed
  to successfully build applications.
* Self implemented "new" operators and purecall stub are not needed.

The main "makeall" script excepts the option "--nosample" now, which
suppresses building the latest Circle sample, which will fail with
STDLIB_SUPPORT defined.

Applications will be linked using libgcc.a now in any case. Own
implementations of the divide routines are not used any more.