CMemorySystem can be instantiated more than once
Normally there is only one instance of class CMemorySystem in Circle.
But with the recent changes for STDLIB_SUPPORT it's possible that
programs (using old samples as template) instantiate CMemorySystem
in CKernel, which has already created in sysinit() before calling
the static constructors. This would cause the system initialization
to fail without notice and is difficult to debug. Because of this
CMemorySystem can be instantiated more than once now.