FIXED: Mouse cursor will be updated from TASK_LEVEL now
Updating the mouse cursor requires a property mailbox call, which
can only be done safely from TASK_LEVEL. Previously this was done
automatically from IRQ_LEVEL in class CMouseBehaviour, which is
in fact not allowed. Now an application, which uses the mouse cursor
(cooked mode), has to call frequently CUSBMouseDevice::UpdateCursor()
from TASK_LEVEL, so that the cursor can move on the screen. The mouse
driver and samples (including addon/ugui) have been updated that way.