Initial attempt to handle floating point exceptions
Circle would normally need a "VFP support code" module to handle
specific floating point conditions. Because this cannot be quickly
added as a first attempt FP exceptions will be detected and ignored.
Additionally the "Default NaN mode" has been activated on the FPU,
so that invalid conditions lead to a NaN result.