CTCPConnection::Receive() returns error on FIN from pear
When a TCP connection received a FIN (disconnect) and the application
called CSocket::Receive() in the resulting TCP CLOSE-WAIT state, it did
not get an error so far. Because of this, the disconnect has not been
detectable. Now an error will be returned in this case. This is a
change in the behaviour of the CSocket API and may cause problems with
existing applications using TCP. You may have to check this.

class CHTTPClient updated to handle this API change.