Preparations for new USB gamepad drivers
Checking the HID interface class too in CUSBHIDDevice::Configure()
before switching to boot protocol. This allows using the class
CUSBHIDDevice for USB devices with non-HID-class interfaces.

CUSBFunction::SelectInterfaceByClass() added. By default USB devices
which have been detected using the vendor and product ID will be
controlled using the first interface descriptor in the USB
configuration. With this method, which has to be called in the
constructor of the driver class, an other interface descriptor
can be used.