CUSBGamePadPS3Device based on CUSBGamePadStandardDevice now
class CUSBGamePadPS3Device uses the HID descriptor parser in
CUSBGamePadStandardDevice to determine the actual size of the
HID report, because there are PS3 gamepads with different report
sizes. Once the report size is known, further received reports
will not be parsed by the HID descriptor parser any more. Because
the report format is known then, it can be handled in a quicker
decode routine.

The special support for the PS3 gamepad in CUSBGamePadStandardDevice
itself has been removed.