FIXED: Wrong stack alignment on exception in some cases
Stack pointer must be 8 bytes aligned, when calling out to a C-
function from an exception stub. This has not been the case with
SAVE_VFP_REGS_ON_IRQ/FIQ defined and for abort stubs. Problems
caused by this issue are not known, but to be sure it has been
corrected now.