Prevent usage of unconstructed static spin locks
Class CBcmMailBox is used early before the static constructors have
been called, but it is using a static spin lock. To prevent usage of
an unconstructed object, CBcmMailBox::CBcmMailBox() gets a boolean
bEarlyUse now, which allows to use the mail box without spin lock.
This is not a problem early, because there is only one thread of
execution at this time.