Automatic GPIO clock source and divider selection added
The clock rates of the GPIO clock sources of the RPi 4 are different
from RPi 1-3. That's why the static selection of clock sources and
dividers does not work any more. The new method CGPIOClock::StartRate()
automatically selects an available clock source and calculates the
divider from a given target clock rate. The divider must be an integer
value for now (MASH is not used). Users of the class CGPIOClock have
been updated to use this feature.