FIQ with GIC-400 supported (AArch32 only)
A Secure Monitor is used to do secure accesses to the GIC-400, which
is required to initialize group 0 interrupts to trigger the FIQ. This
cannot be implemented in AArch64, because it is unknown there, where
the MVBAR register points to. So returning to EL3 is impossible and
FIQ cannot be initialized and used.