class CXHCIDevice and alias CUSBHCIDevice added
class CXHCIDevice will be the USB host controller interface driver for
the USB xHCI device of the RPi 4 one day. It's incomplete at the moment
and fakes an USB without attached devices. CUSBHCIDevice is a generic
alias name for the USB host controller interface driver class, which is
used in the target system (CDWHCIDevice or CXHCIDevice). CUSBHCIDevice
is used (instead of CDWHCIDevice) in all applications now, which need
USB support. All Circle samples have been updated that way.