addon/SDCard/ library works on Raspberry Pi 4 now
Currently the legacy EMMC1 controller is used. CardReset() may fail
while executing the SEND_SCR command with DATA_TIMEOUT for unknown
reason. As workaround the whole CardReset() is retried. This worked
well in a longer test and later problems did not occur.

CBcmPropertyTags::GetTag() did report an error, if the response bit
was not set in the value length header field. This did not work with
the PROPTAG_SET_SET_GPIO_STATE function and has been changed.