tools/bootloader requests core clock rate from firmware
The core clock rate, which is needed to calculate the baud rate divider
for the Mini-UART, was fixed to 250 MHz before. With a recent firmware
change this does not work any more, because the core clock rate varies
for the different Raspberry Pi models. The divider is calculated from
the actual core clock rate now.