Network performance vastly improved on RPi 4 AArch64
Network performance was very bad on RPi 4 with AArch64 since it was
supported in Step40 (e.g. RX bandwidth only 5 Mbps with iperf on Fast
Ethernet). The Genet Gigabit Ethernet controller provides five TX
queues with different priorities. Because Circle does not support TX
priority mapping, only one queue is used for all frame transmits.
Previously the lowest priority TX queue was selected, because it has
more buffers (128) than the other queues (32). As is turns out this
decision was wrong, because if a connected host transmitted frames to
us with high-speed, our TCP ACKs were only be send delayed. Hence the
transfer often delayed because of many re-transmits.

Now the highest priority TX queue is used, which solves the problem,
but because of the reduced queue size frame drops occurred in
CBcm54213Device::SendFrame(). A method IsSendFrameAdvisable() is
implemented which allows to check, if send is possible at a specific
time or if the arrived TX frames should stay queued in main memory.
CNetDeviceLayer::Process() is updated to achieve this. The RX bandwidth
is 75 Mbps with iperf now.