Construct CMemorySystem object early in sysinit() in any case
Previously a CMemorySystem object had to be instanced as first member of CKernel
with STDLIB_SUPPORT < 2 to be able to select MMU usage by the application. This
caused several problems. For instance newer compilers may generate code in the
early stages, which does not work without MMU. Also constructors of static
objects were called with MMU off before.

Now that the CMemorySystem object is constructed early, these problems should be
solved, but all applications have to run with MMU on. If the application
constructs a second instance of CMemorySystem in CKernel, this is not a problem.
It is used as an alias for the first instance.

Update sample/02-screenpixel and sample/03-screentext to consider, that MMU is
always on.