May manually acknowledge GPIO event detect status
Previously in the classes CGPIOPin and CGPIOPinFIQ the GPIO event
detect status was automatically acknowledged, when an interrupt
occurred. Now the user can decide to acknowledge the event detect
status manually from its own interrupt handler, where this operation
fits best. There is a new bAutoAck parameter for ConnectInterrupt()
and an AcknowledgeInterrupt() method.