usb/usbhostcontroller: UpdatePlugAndPlay() returns status
If CUSBHostController::UpdatePlugAndPlay() returned FALSE, there is
no update to the USB device tree and nothing needs to be done to react
on a device tree update. This can reduce the effort, which is needed
to poll for devices to appear. The PnP enabled samples have been
updated that way.