usb/xhci: Support Raspberry Pi 400
With newer firmware versions the PCIe inbound memory window is not the
same for all Raspberry Pi 4 models any more. This applies to the
Raspberry Pi 400 too. This memory window is now determined from the
Devicetree blob, provided by the firmware, configured in the PCIe host
bridge and used by the xHCI driver.

Issue #166