FIXED: usb: Number of digital buttons for known gamepads
The number of digital buttons in the gamepad state (nbuttons) was set
to the number of physical available "slots" in the gamepad report for
known gamepads too. This was wrong, because the digital buttons are
remapped to a standard mapping with more possible buttons for known
gamepads and an application, which recognizes only this number of
buttons, would ignore some buttons.

Issue #201