usb/usbmassdevice: Fix hard errors in reset recovery
The recovery function did not work at all, because it was never used
before. Now on the RPi 3A+, Zero and probably A+ it is necessary and
has been corrected. Unfortunately this does not solve all problems
with USB flash drive access on these RPi models. Reads or writes may
still fail, but not as often.

Issue #211