soundbasedevice: Add Read() and related methods
There has been an inconsistency with SoundFormatSigned24, because there
are two different variants of it: one 24-bit format, which occupies 3
bytes per sample and one, which occupies 4 bytes per sample (value of
one byte ignored). Previously the 4-bytes-format was used for
CI2SSoundBaseDevice::Write(), if SoundFormatSigned24 was selected as
write format with 2 channels. For all other sound devices the 3-bytes-
format was used. To fix this, a new SoundFormatSigned24_32 is
introduced, which is the 4-bytes-format. SoundFormatSigned24 is now
the 3-byte format in any case. In existing applications, which use
CI2SSoundBaseDevice::Write(), it may be necessary to change the
parameter of SetWriteFormat() to SoundFormatSigned24_32.