Improve energy efficiency of AArch64 lock acquire
The 32-bit implementation of the spinlock Acquire() function uses a
"wfe" instruction to send the core to sleep if it can't acquire a lock,
and the corresponding Release() function will send an event with the
"sev" instruction to wake up waiting cores.

Add this sleep functionality to the 64-bit implementation to improve
energy efficiency in AArch64 builds.

N.B. a "sev" instruction is not required for AArch64 Release() because
an event is automatically generated (see section B2.9.6 of "Arm
Architecture Reference Manual Armv8, for Armv8-A architecture profile").