classallocator: Store extends when inited more than once
If there is a class, which can be instantiated more than once, and
which uses another class/struct internally, the class-specific
allocator for that class/struct will be initialized more than once
too. This was forbidden before. Now it is allowed for the protected
allocator. The class store will be extended then by the given number
of objects.