USB sound supported for Raspberry Pi 4 only
Unfortunately it was not possible to support USB audio streaming
reliable for Raspberry Pi 1-3 and Zero also for output so far. The
required support for isochronous USB endpoints for the DWHCI USB host
controller is still very fragile. Because there was no Circle release
for half a year, and there should be one soon, the decision has been
made to restrict the USB audio streaming support on the Raspberry Pi 4
and compatible models with xHCI USB host controller.