usb/usbaudiostreaming: Workaround for RME Babyface Pro
The RME Babyface Pro references a source entity #2 for the speaker
output terminal, but this entity does not exist. The corresponding USB
streaming input terminal has the identifier #3. Because of this, it was
not possible to find the corresponding terminal type for the input
terminal and additionally an assertion in usbaudiofunctopology.cpp
failed. This workaround should fix the problem.

See: #336