FIXED: usb/xhciendpoint: Isoch frame number synchronization
The modulo comparison of frame numbers in GetIsoFrameID_SIA() was wrong,
so that m_usLastIsoFrameIndex was never reset to usMinIndex. This could
lead to many Missed Service and Ring Over/Underrun Errors and an Event
Ring Full Error, when isochronous transfers were started. An "in sync"
status is maintained now, which reflects the need to synchronize the
frame numbers (again).