usb/usbaudiostreaming: Add workaround for Dirtywave M8
There seem to be issues with the isochronous feedback endpoint of this
device. Under some conditions it sends empty or invalid responses. With
this workaround the isochronous output data endpoint is used in
synchronous mode, instead of in asynchronous mode with feedback. It has
to be tested again, if this works reliable, and if this is workaround
is really necessary.