Enable strict alignment, where necessary
GCC 12 may generate unaligned memory accesses, which is normally not
a problem, but can be, when un-cached (coherent) memory is accessed,
or in the early stages, when the MMU is not initialized yet. This
affects the following libraries:

* lib/libcircle.a
* lib/usb/libusb.a (xHCI, Raspberry Pi 4 only)
* addon/vc4/vchiq/libvchiq.a

lib/input/libinput.a is not affected. It uses coherent memory in
rpitouchscreen.o, but it is safe to do only byte-wide accesses and
uses memcpy(), which is safe to be strict aligned too.