ptrlist: Make some methods const
There were some methods in CPtrList, which do not modify class members,
but were not declared const. This prevented using this methods in const
methods of other classes and thus the whole class. Declare these methods
as const, which makes more sense and has not influence of existing class
users. GetPtr() can be declared static, because it does not access class
members at all.

Document CPtrList for Doxygen.