addon/Properties: Support sections
The property files can have section headers (like "[section-name]") now,
which group a number of property pairs together. Prior to getting and
setting a property, the method SelectSection() can be called to select
the section name for the following accesses. By default the
DefaultSection is accessed, which has no section header and is located
at the beginning of a property file. If SelectSection() is not called,
the class behaves like before. The sample program is updated.

See: #409