rp1: Can be instantiated multiple times
The CRP1 class must be created and initialized prior to several other
devices, but existing applications are not prepared for this. Especially
the xHCI requires the CRP1. We add the CRP1 to the xHCI driver and if it
was already instantiated before, we ignore the second instance. So
existing USB applications need not be modified, and newer applications,
which explicitly add the CRP1 to CKernel (e.g. for using GPIO pins) do
work either.