addon/lvgl: Add LVGL mouse cursor for Raspberry Pi 5
The Raspberry Pi 5 firmware does not support the previously available
mouse cursor. Setup a LVGL software mouse cursor instead. The class
CMouseBehaviour can still be used to maintain the position of the mouse
pointer on the display. The firmware access for the mouse cursor has
been disabled for Raspberry Pi 5.