sysinit: Instantiate CSouthbridge early
Many programs will need a CSouthbridge object, often set-up before
other classes (e.g. CGPIOPin, CSerialDevice) can be instantiated.
Therefore a CSouthbridge object is created and initialized early in
sysinit(). Because CSouthbridge needs CInterruptSystem, it is created
and initialized before. Because there is a second instance of
CInterruptSystem then in existing applications, this second instance
does nothing and its methods simply call the methods of the early
created instance. Because error messages cannot be generated that early,
the Act LED is used to show an error blink pattern.