usb/xhci: Do not use USE_XHCI_INTERNAL for Raspberry Pi 5
The (DWC3) USB controllers of the Raspberry Pi 5 are not internal in
the SoC, so using this option was misleading. It implied xHCI v1.10
instead of v1.00 and the CSZ (64-bit context size) option active. There
is a new XHCI_CONTEXT_SIZE define now. The inbound memory window is read
from the DTB now. Its default setting was wrong and did only work,
because XHCI_TO_DMA() was a NOP with USE_XHCI_INTERNAL.