FIXED: usb/dwhci: SetCompleteOnNAK() in no-hub configurations
USB MIDI devices respond with NAK on its Bulk-in EP, when there is
no data available. This shall complete the active USB request, which was
working in configurations with an USB hub involved (split transfers),
but not, when the USB device was directly connected to the root port.
Fix it.

Issue #467