usb/usbcdcethernet: Support RTL8152 adapter
Before the USB descriptor parser was working on a local copy of the
configuration descriptor. This was leading to the issue, that not the
right data interface was set with the RTL8152 adapter and data transfer
was impossible. Add the Realtek RTL8152 adapter to the devices with
multiple configuration descriptors, where the second descriptor has
to be used.