FIXED: machineinfo: Deleted DTB in secondary instance
The class CMachineInfo can be instantiated multiple times and the
secondary instances will be directed to the primary one, which is
instantiated in sysinit. This fixes the issue, that the DTB member
was deleted, which a secondary instance was destroyed, so it was not
available any more for other instances.