2dgraphics: Speed up operation
Before a new graphics frame has been prepared in the uncached display
buffer. Now a separate cached buffer is used, which speeds up operation.
The separate buffer is copied to the display buffer in the vertical sync
phase to suppress flickering. Now this class can be used on the
Raspberry Pi 5 too, but without vertical sync support.

Issue #489