machineinfo: Add Raspberry Pi 500 and CM5 (Lite)
These Raspberry Pi models are not fully supported and not tested with
Circle, because they were not available for this purpose. They will be
added for completeness. It is likely, that they work to some degree
with Circle, but some peripherals will not work.

boot/Makefile: Add related DTB files