net/mdnsdaemon: Add class CmDNSDaemon
This mDNS responder task determines and maintains our mDNS hostname on
the local network. Name collisions with other hosts will be resolved by
appending a numeric suffix to the hostname. The daemon is automatically
started, when someone requests a pointer to it. This has been added to
the CmDNSPublisher class, which works in conjunction with CmDNSDaemon
now. The test/mdns-publisher uses the numeric suffix to generate an
unique service name.

See: #564