FIXED: net/mdnsdaemon: ParseResponse() did not terminate
When ParseResponse() did not found a valid "A" resource record in
a message, it might have continued to parse arbitrary data in memory.
This might have allocated large memory blocks and led to an exhausted
heap memory under some circumstances. Also the hostname in the received
response was not compared with our own hostname to ensure, that the
message was to us.

Issue #607