FIXED: net/tcpconnection: Wrong SND./RCV.NXT handling on close
The SND./RCV.NXT variables were increased on each send or receive of a
FIN segment. Now with the speeded-up net buffer handling this caused an
invalid termination of connection after close from the application. FIN
was repeated with an increasing sequence number and received FIN
segments were not accepted. The connection was closed after RST from
peer. Fix it.