net/socket: Add MSG_MORE send flag
Can specify MSG_MORE as flag for Send[To](). This will remove the PSH
flag from the TCP data segment to speed up the transfer. MSG_MORE is
automatically applied, when data with a length of over MSS is send, for
all segments, but the last one. With MSG_MORE even the last segment is
not pushed to the application. On receive PSH is automatically applied,
when the RX queue contains at least 64K bytes. UDP ignores this flag.

Discard net buffers in the socket layer, which could not be sent by the
transport layer.