net/tcpconnection: Add reassembly queue
This speeds up receiving on a Gigabit Ethernet e.g. on the RPi 1 very
much, were the Ethernet switch has to drop packets, because the RPi 1
has only 100BaseT Ethernet. This leads to Fast Retransmits from the
sender, which were handled disadvantageous. Overlapping TCP data
segments are currently not supported, so this will be disabled, when
they are detected.